employee

+++ BE/dacn/src/main/java/com/hrm/dacn/repositories/EmployeeRepository.java
@@ -4,8 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.hrm.dacn.entities.Employee;
 
-import jakarta.persistence.criteria.CriteriaBuilder.In;
-
-public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
+public interface EmployeeRepository extends JpaRepository<Employee, Long> {
 
 }
